refactor: remove folder redundancy
