+++ cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/FelipeMCassiano/gorvus/internal/commands"
+
+func main() {
+	commands.Execute()
+}
