--- gorvus/cmd/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/FelipeMCassiano/gorvus/gorvus/internal/commands"
-
-func main() {
-	commands.Execute()
-}
